Implement Ceiling distribution

+++ distribution_tools/praise/forum_post_generator.py
@@ -7,7 +7,7 @@ def generate_post(params, ROOT_INPUT_PATH=""):
     token_table_path = ROOT_INPUT_PATH + \
         "distribution_results/raw_csv_exports/final_praise_token_allocation.csv"
     params["token_allocation_per_reward_system"] = list(
-        map(int, params["token_allocation_per_reward_system"]))
+        map(float, params["token_allocation_per_reward_system"]))
 
     data = pd.read_csv(praise_path)
     start_date = pd.to_datetime(data["DATE"].min()).date()
